update curd, search, filter task

+++ src/main/java/com/example/demo/controller/TasksController.java
@@ -8,10 +8,10 @@ import com.example.demo.service.TasksService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity; 
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*; 
 
-import java.util.List; 
+import java.util.List;
 
 
 @RestController
@@ -38,15 +38,34 @@ public class TasksController {
     public ResponseEntity<Tasks> createTask(@RequestBody TasksDto task) {
         return new ResponseEntity<Tasks>(taskService.createNewTask(task),HttpStatus.CREATED); 
     } 
+
     @PutMapping("/{id}") 
-    public ResponseEntity<TasksDto> updateTask(@PathVariable Long id, @RequestBody TasksDto task) { 
-        taskService.updateTask(task, id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<Tasks> updateTask(@PathVariable Long id, @RequestBody TasksDto task) { 
+        if (taskService.checkAuthor(id)){
+            return new ResponseEntity<Tasks>(taskService.updateTask(task, id), HttpStatus.OK);
+        }else {
+            return new ResponseEntity<Tasks>(taskService.findTaskById(id), HttpStatus.BAD_REQUEST); 
+        }
+        
     } 
     @DeleteMapping("/{id}") 
     public ResponseEntity<Boolean> getAllTasks(@PathVariable Long id) { 
-        taskService.deleteTask(id); 
-        return ResponseEntity.ok(true); 
+        if (taskService.checkAuthor(id)){
+            taskService.deleteTask(id);
+            return ResponseEntity.ok(true); 
+        }else {
+            return new ResponseEntity<>(false, HttpStatus.BAD_REQUEST); 
+        }
+    } 
+
+    @PostMapping("/search") 
+    public ResponseEntity<List<Tasks>> seachByTaskName(@RequestParam String keyWord) {
+        List<Tasks> listtTasks = taskService.searchByTaskName(keyWord);
+        if (listtTasks.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(listtTasks, HttpStatus.OK);
+        }
     } 
 }
